AutoCompleter's own class
